Delay one-wire initialization
This pulls out the creation of the OneWire and DallasTemperature instances from the EC class. This may not be entirely necessary, but it looked like the Arduino was hanging when I tried to set these up in the EC constructor. That very likely was the call to “begin”. I’ve moved the “begin” call into a separate EC method which we now call in setup().

Note that this is a WIP. I’ll need actual sensors before I can test, so I’m sending dummy data for now.