Update code to use PH and EC readings
We now read EC and PH continuously. When a request is made for values, we simpy read the current value in those instances. There is a chance of readings changing as we read them, but changes should be small anyway, so I think that is a concern (famous last words?)